AI Chat Bot KRMU/ai_engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai==1.51.0
langchain==0.3.27
faiss-cpu==1.13.2
pypdf==5.0.1
numpy==2.1.2
python-dotenv==1.0.1